--- app/components/footer.vue
@@ -1,4 +1,0 @@
-<template>
-    <div>Text</div>
-</template>
-<script setup></script>
