+++ app/components/header.vue
@@ -1,7 +1,27 @@
 <template>
-    <div class="text-3xl font-bold underline">
-        Text
-    </div>
-    <div></div>
+  <header class="p-4 bg-gray-100">
+    <nav class="flex gap-4">
+      <div v-if="!authStore.isAuthenticated">
+          <NuxtLink to="/login">
+            Login
+          </NuxtLink>
+          <NuxtLink to="/signup">
+            Signup
+          </NuxtLink>
+      </div>
+      <div v-else>
+          <NuxtLink to="/user" class="text-blue-500 hover:text-blue-600">
+            User
+          </NuxtLink>
+          <NuxtLink to="/note" class="text-blue-500 hover:text-blue-600">
+            Note
+          </NuxtLink>
+      </div>
+    </nav>
+  </header>
 </template>
-<script setup></script>
+<script setup>
+import { useAuthStore } from "~/stores/auth";
+const authStore = useAuthStore();
+console.log(authStore.isAuthenticated);
+</script>
