+++ app/layouts/default.vue
@@ -2,7 +2,6 @@
   <div>
     <Header />
     <slot />
-    <Footer />
   </div>
 </template>
 <script setup></script>
