+++ app/pages/note/index.vue
