+++ app/pages/user/index.vue
@@ -1,0 +1,2 @@
+<template></template>
+<script setup></script>
